Fixes reference transform IncludeNamespace
Dismiss namespaces from PrefixList that are not used within target node. Default PHP behaviour of the DOMNode::C14N method to canonicalize xml is to add missing namespace definitions from the root node to the element being canonicalized. This causes the element to be different at the time of digest calculation, since the missing namespace(s) will be added. This causes digest validation to fail. This fix ensures only namespace(s) are added to the element that are actually used within the element as child nodes.